Move TopologyManager TestPolicyMerge() to shared test file
